requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.1.0
beautifulsoup4>=4.12.3
# Указываем диапазон httpx, который нужен другим пакетам и совместим с новым telegram-bot
httpx>=0.27,<1
mcp>=0.2.1
langchain-mcp-adapters>=0.1.6
langchain>=0.1.16
langchain_community>=0.0.34
langchain-mcp-adapters>=0.1.6
# Другие потенциально необходимые зависимости (версии могут быть скорректированы)
langchain>=0.1.16
langchain_community>=0.0.34
# Оставьте эти ограничения более гибкими, если они вызывают проблемы
# anyio>=3.7,<5 # Часто управляется другими пакетами
# numpy>=1.21,<1.25 # Может вызывать конфликты, проверьте необходимость
# pydantic>=2.5,<3.0 # Часто управляется другими пакетами
# pydantic-settings>=2.0.3 # Часто управляется другими пакетами